+++ RandBLAS/dense.hh
@@ -16,41 +16,6 @@
 #include <typeinfo>
 
 
-/*
-Paradigm for APIs involving structs:
-    Free-functions when there's no memory to manage
-    Member functions when there IS memory to manage, or in initializing.
-        We want to make this library hard to misuse in C++.
-    We provide APIs that we require people use to ensure that structs are
-        in a valid state. If you want to initialize the struct yourself
-        we won't stop you, but we also take no responsibility for the 
-        inevitable segfaults.
-
-TODO: have a discussion around using smart pointers for memory safety.
-    Burlen thinks we should seriously consider using smart pointers.
-*/
-
-/*
-Currently have non-deterministic behavior (tests pass sometimes, fail sometimes).
-I suspect there's some memory management mistake leading to undefined behavior.
-
-      Start 18: TestDenseMoments.Gaussian
-    18/34 Test #18: TestDenseMoments.Gaussian ....................***Failed    0.11 sec
-    Running main() from /tmp/googletest-20220910-45435-1kz3pjx/googletest-release-1.12.1/googletest/src/gtest_main.cc
-    Note: Google Test filter = TestDenseMoments.Gaussian
-    [==========] Running 1 test from 1 test suite.
-    [----------] Global test environment set-up.
-    [----------] 1 test from TestDenseMoments
-    [ RUN      ] TestDenseMoments.Gaussian
-    /Users/riley/BALLISTIC_RNLA/randla/RandBLAS/test/src/test_dense.cc:48: Failure
-    The difference between mean and 0.0 is 0.01195285380042985, which exceeds 1e-2, where
-    mean evaluates to -0.01195285380042985,
-    0.0 evaluates to 0, and
-    1e-2 evaluates to 0.01.
-    [  FAILED  ] TestDenseMoments.Gaussian (112 ms)
-    [----------] 1 test from TestDenseMoments (112 ms total)
-*/
-
 namespace RandBLAS::dense {
 
 using namespace RandBLAS::base;
@@ -61,24 +26,32 @@ using namespace RandBLAS::base;
 /// RandBLAS currently have i.i.d. entries. This enumeration specifies
 /// the distribution of the entries of such a sketching operator.
 enum class DenseDistName : char {
-
     // ---------------------------------------------------------------------------
     ///  Gaussian distribution with mean 0 and standard deviation 1
     Gaussian = 'G',
+
     // ---------------------------------------------------------------------------
     ///  uniform distribution over [-1, 1].
-    Uniform = 'U'
-    
-    //// ---------------------------------------------------------------------------
-    ///  uniform distribution over \math{\{-1, 1\}}.
-    // Rademacher = 'R',
-    //// ---------------------------------------------------------------------------
-    ///  A flag that a sketching operator's distribution should be 
-    ///  uniform over row-orthonormal or column-orthonormal matrices.
-    // Haar = 'H',
-    
-    //DisjointIntervals = 'I' // might require additional metadata.
+    Uniform = 'U',
+
+    // ---------------------------------------------------------------------------
+    ///  entries are defined only by a user-provided buffer
+    BlackBox = 'B'
 };
+
+enum class FillAxis : char {
+    // ---------------------------------------------------------------------------
+    ///  short-axis vectors (cols of a wide matrix, rows of a tall matrix)
+    Short = 'S',
+
+    // ---------------------------------------------------------------------------
+    ///  long-axis vectors (rows of a wide matrix, cols of a tall matrix)
+    Long = 'U'
+};
+
+// FillAxis swap_axis(FillAxis fa) {
+//     return (fa == FillAxis::Long) ? FillAxis::Short : FillAxis::Long;
+// }
 
 // =============================================================================
 /// A distribution over dense sketching operators.
@@ -95,6 +68,10 @@ struct DenseDist {
     // ---------------------------------------------------------------------------
     ///  The distribution used for the entries of the sketching operator.
     const DenseDistName family = DenseDistName::Gaussian;
+
+    // ---------------------------------------------------------------------------
+    ///  The order in which the buffer should be populated, if sampling iid.
+    const FillAxis fill_axis = FillAxis::Long;
 };
 
 // =============================================================================
@@ -185,6 +162,8 @@ DenseSkOp<T,RNG>::DenseSkOp(
 {   // sanity checks
     randblas_require(this->dist.n_rows > 0);
     randblas_require(this->dist.n_cols > 0);
+    if (dist_.family == DenseDistName::BlackBox)
+        randblas_require(this->buff != nullptr);
 }
 
 template <typename T, typename RNG>
@@ -193,9 +172,6 @@ DenseSkOp<T,RNG>::~DenseSkOp() {
         delete [] this->buff;
     }
 }
-
-
-
 
 
 
@@ -219,12 +195,23 @@ DenseSkOp<T,RNG>::~DenseSkOp() {
  * @returns the updated CBRNG state
  */
 template <typename T, typename RNG, typename OP>
-auto fill_rmat(
+RNGState<RNG> fill_rmat(
     int64_t n_rows,
     int64_t n_cols,
     T* mat,
-    const RNGState<RNG> & seed
+    const RNGState<RNG> & seed,
+    FillAxis fa = FillAxis::Long
 ) {
+    //  Assume that this function fills column-major by default.
+    //      If this matrix is tall and fa==Long, then we continue as normal.
+    //      If this matrix is wide and fa==short, then we continue as normal.
+    //      If we're tall and fa=short, then we should fill the wide transpose.
+    //      If we're wide and fa=long, then we should fill the tall transpose.
+    if (fa == FillAxis::Long && n_cols > n_rows)
+        return fill_rmat<T,RNG,OP>(n_cols, n_rows, mat, seed, fa);
+    if (fa == FillAxis::Short && n_rows > n_cols)
+        return fill_rmat<T,RNG,OP>(n_cols, n_rows, mat, seed, fa);
+
     RNG rng;
     // clang chokes on this w/ omp due to internal use of lambdas, fixed in C++20
     //auto [c, k] = seed;
@@ -297,17 +284,11 @@ auto fill_buff(
 ) {
     switch (D.family) {
         case DenseDistName::Gaussian:
-            return fill_rmat<T,RNG,r123ext::boxmul>(D.n_rows, D.n_cols, buff, state);
+            return fill_rmat<T,RNG,r123ext::boxmul>(D.n_rows, D.n_cols, buff, state, D.fill_axis);
         case DenseDistName::Uniform:
-            return fill_rmat<T,RNG,r123ext::uneg11>(D.n_rows, D.n_cols, buff, state);
-        //case DenseDistName::Rademacher:
-        //    throw std::runtime_error(std::string("Not implemented."));
-        //case DenseDistName::Haar:
-            // This won't be filled IID, but a Householder representation
-            // of a column-orthonormal matrix Q can be stored in the lower
-            // triangle of Q (with "tau" on the diagonal). So the size of
-            // buff will still be D.n_rows*D.n_cols.
-        //    throw std::runtime_error(std::string("Not implemented."));
+            return fill_rmat<T,RNG,r123ext::uneg11>(D.n_rows, D.n_cols, buff, state, D.fill_axis);
+        case DenseDistName::BlackBox:
+            throw std::invalid_argument(std::string("fill_buff cannot be called with the BlackBox distribution."));
         default:
             throw std::runtime_error(std::string("Unrecognized distribution."));
     }
@@ -455,6 +436,13 @@ void lskge3(
     int64_t ldb
 ){
     randblas_require(S0.layout == layout);
+    if (S0.layout != layout) {
+        // I think it suffices to change this flag, and we can proceed with dimensions unchanged.
+        // we might need to change LDS though ...
+        transS = (transS == blas::Op::NoTrans) ? blas::Op::Trans : blas::Op::NoTrans;
+        // what about computing ptr_offset? Do we compute that w.r.t S0.layout, or plain layout?
+    }
+    
 
     DenseSkOp<T,RNG> S0_shallow_copy(S0.dist, S0.seed_state, S0.buff, S0.layout);
     T *S0_ptr = S0.buff;
@@ -490,16 +478,20 @@ void lskge3(
 
     // Sanity checks on dimensions and strides
     int64_t lds, pos;
-    if (layout == blas::Layout::ColMajor) {
+    if (S0.layout == blas::Layout::ColMajor) {
         lds = S0.dist.n_rows;
-        pos = i_os + lds * j_os;
         randblas_require(lds >= rows_submat_S);
+        pos = i_os + lds * j_os;
+    } else {
+        lds = S0.dist.n_cols;
+        randblas_require(lds >= cols_submat_S);
+        pos = i_os * lds + j_os;
+    }
+
+    if (layout == blas::Layout::ColMajor) {
         randblas_require(lda >= rows_A);
         randblas_require(ldb >= d);
     } else {
-        lds = S0.dist.n_cols;
-        pos = i_os * lds + j_os;
-        randblas_require(lds >= cols_submat_S);
         randblas_require(lda >= cols_A);
         randblas_require(ldb >= n);
     }
