+++ RandBLAS/test_util.hh
@@ -6,7 +6,49 @@
 #include <numeric>
 #include <iostream>
 
+#include <type_traits>
+#include <typeinfo>
+#ifndef _MSC_VER
+#   include <cxxabi.h>
+#endif
+#include <memory>
+#include <string>
+#include <cstdlib>
+
+#include <RandBLAS/dense.hh>
+#include <RandBLAS/sparse.hh>
+#include <RandBLAS/ramm.hh>
+#include <RandBLAS/util.hh>
+#include <math.h>
+
+
 namespace RandBLAS_Testing::Util {
+
+template <class T>
+std::string type_name() { // call as type_name<obj>()
+    typedef typename std::remove_reference<T>::type TR;
+    std::unique_ptr<char, void(*)(void*)> own
+           (
+#ifndef _MSC_VER
+                abi::__cxa_demangle(typeid(TR).name(), nullptr,
+                                           nullptr, nullptr),
+#else
+                nullptr,
+#endif
+                std::free
+           );
+    std::string r = own != nullptr ? own.get() : typeid(TR).name();
+    if (std::is_const<TR>::value)
+        r += " const";
+    if (std::is_volatile<TR>::value)
+        r += " volatile";
+    if (std::is_lvalue_reference<T>::value)
+        r += "&";
+    else if (std::is_rvalue_reference<T>::value)
+        r += "&&";
+    return r;
+}
+
 
 /** Tests two floating point numbers for approximate equality.
  * See https://randomascii.wordpress.com/2012/02/25/comparing-floating-point-numbers-2012-edition/
@@ -155,6 +197,209 @@ void matrices_approx_equal(
     }
 }
 
-} // end namespace RandBLAS_Testing::Util
+template <typename T>
+void sparseskop_to_dense(
+    RandBLAS::sparse::SparseSkOp<T> &S0,
+    T *mat,
+    blas::Layout layout,
+    bool take_abs = false
+) {
+    RandBLAS::sparse::SparseDist D = S0.dist;
+    for (int64_t i = 0; i < D.n_rows * D.n_cols; ++i)
+        mat[i] = 0.0;
+    auto idx = [D, layout](int64_t i, int64_t j) {
+        return  (layout == blas::Layout::ColMajor) ? (i + j*D.n_rows) : (j + i*D.n_cols);
+    };
+    int64_t nnz;
+    if (D.family == RandBLAS::sparse::SparsityPattern::SASO) {
+        nnz = D.vec_nnz * MAX(D.n_rows, D.n_cols);
+    } else {
+        nnz = D.vec_nnz * MIN(D.n_rows, D.n_cols);
+    }
+    for (int64_t i = 0; i < nnz; ++i) {
+        int64_t row = S0.rows[i];
+        int64_t col = S0.cols[i];
+        T val = S0.vals[i];
+        if (take_abs)
+            val = abs(val);
+        mat[idx(row, col)] = val;
+    }
+}
 
+template <typename T, typename RNG>
+void reference_lskges(
+    blas::Layout layout,
+    blas::Op transS,
+    blas::Op transA,
+    int64_t d, // mat(B) is d-by-n
+    int64_t n, // op(mat(A)) is m-by-n
+    int64_t m, // op(submat(S)) is d-by-m
+    T alpha,
+    RandBLAS::sparse::SparseSkOp<T,RNG> &S,
+    int64_t i_os,
+    int64_t j_os,
+    const T *A,
+    int64_t lda,
+    T beta,
+    T *B,  // expected value produced by LSKGES; compute via GEMM.
+    T *E,  // allowable floating point error; apply theory + compute by GEMM.
+    int64_t ldb
+){
+    randblas_require(d > 0);
+    randblas_require(m > 0);
+    randblas_require(n > 0);
+    std::vector<T> S_dense(S.dist.n_rows * S.dist.n_cols);
+    sparseskop_to_dense<T>(S, S_dense.data(), layout, false);
+    std::vector<T> S_dense_abs(S.dist.n_rows * S.dist.n_cols);
+    sparseskop_to_dense<T>(S, S_dense_abs.data(), layout, true);
+
+    // Dimensions of mat(A), rather than op(mat(A))
+    int64_t rows_mat_A, cols_mat_A, rows_submat_S, cols_submat_S;
+    if (transA == blas::Op::NoTrans) {
+        rows_mat_A = m;
+        cols_mat_A = n;
+    } else {
+        rows_mat_A = n;
+        cols_mat_A = m;
+    }
+    // Dimensions of submat(S), rather than op(submat(S))
+    if (transS == blas::Op::NoTrans) {
+        rows_submat_S = d;
+        cols_submat_S = m;
+    } else {
+        rows_submat_S = m;
+        cols_submat_S = d;
+    }
+    // Sanity checks on dimensions and strides
+    int64_t lds, pos, size_A, size_B;
+    if (layout == blas::Layout::ColMajor) {
+        lds = S.dist.n_rows;
+        pos = i_os + lds * j_os;
+        randblas_require(lds >= rows_submat_S);
+        randblas_require(lda >= rows_mat_A);
+        randblas_require(ldb >= d);
+        size_A = lda * (cols_mat_A - 1) + rows_mat_A;;
+        size_B = ldb * (n - 1) + d;
+    } else {
+        lds = S.dist.n_cols;
+        pos = i_os * lds + j_os;
+        randblas_require(lds >= cols_submat_S);
+        randblas_require(lda >= cols_mat_A);
+        randblas_require(ldb >= n);
+        size_A = lda * (rows_mat_A - 1) + cols_mat_A;
+        size_B = ldb * (d - 1) + n;
+    }
+
+    // Compute the reference value
+    T* S_ptr = S_dense.data();
+    blas::gemm(layout, transS, transA, d, n, m,
+        alpha, &S_ptr[pos], lds, A, lda, beta, B, ldb
+    );
+
+    // Compute the matrix needed for componentwise error bounds.
+    std::vector<T> A_abs_vec(size_A);
+    T* A_abs = A_abs_vec.data();
+    for (int64_t i = 0; i < size_A; ++i)
+        A_abs[i] = abs(A[i]);
+    if (beta != 0.0) {
+        for (int64_t i = 0; i < size_B; ++i)
+            E[i] = abs(B[i]);
+    }
+    T eps = std::numeric_limits<T>::epsilon();
+    T err_alpha = (abs(alpha) * m) * (2 * eps);
+    T err_beta = abs(beta) * eps;
+    T* S_abs_ptr = S_dense_abs.data();
+    blas::gemm(layout, transS, transA, d, n, m,
+        err_alpha, &S_abs_ptr[pos], lds, A_abs, lda, err_beta, E, ldb
+    );
+    return;
+}
+
+template <typename T, typename RNG>
+void reference_rskges(
+    blas::Layout layout,
+    blas::Op transA,
+    blas::Op transS,
+    int64_t m, // B is m-by-d
+    int64_t d, // op(S) is n-by-d
+    int64_t n, // op(A) is m-by-n
+    T alpha,
+    const T *A,
+    int64_t lda,
+    RandBLAS::sparse::SparseSkOp<T,RNG> &S0,
+    int64_t i_os,
+    int64_t j_os,
+    T beta,
+    T *B, // expected value produced by LSKGES; compute via GEMM.
+    T *E, // allowable floating point error; apply theory + compute by GEMM.
+    int64_t ldb
+) { 
+    using blas::Layout;
+    using blas::Op;
+    //
+    // Check dimensions of submat(S).
+    //
+    int64_t submat_S_rows, submat_S_cols;
+    if (transS == Op::NoTrans) {
+        submat_S_rows = n;
+        submat_S_cols = d;
+    } else {
+        submat_S_rows = d;
+        submat_S_cols = n;
+    }
+    randblas_require(submat_S_rows <= S0.dist.n_rows);
+    randblas_require(submat_S_cols <= S0.dist.n_cols);
+    //
+    // Check dimensions of mat(A).
+    //
+    int64_t mat_A_rows, mat_A_cols;
+    if (transA == Op::NoTrans) {
+        mat_A_rows = m;
+        mat_A_cols = n;
+    } else {
+        mat_A_rows = n;
+        mat_A_cols = m;
+    }
+    if (layout == Layout::ColMajor) {
+        randblas_require(lda >= mat_A_rows);
+    } else {
+        randblas_require(lda >= mat_A_cols);
+    }
+    //
+    // Compute B = op(A) op(submat(S)) by LSKGES. We start with the identity
+    //
+    //      B^T = op(submat(S))^T op(A)^T
+    //
+    // Then we interchange the operator "op" for op(A) and the operator (*)^T.
+    //
+    //      B^T = op(submat(S))^T op(A^T)
+    //
+    // We tell LSKGES to process (B^T) and (A^T) in the opposite memory layout
+    // compared to the layout for (A, B).
+    // 
+    auto trans_transS = (transS == Op::NoTrans) ? Op::Trans : Op::NoTrans;
+    auto trans_layout = (layout == Layout::ColMajor) ? Layout::RowMajor : Layout::ColMajor;
+    reference_lskges(
+        trans_layout, trans_transS, transA,
+        d, m, n, alpha, S0, i_os, j_os, A, lda, beta, B, E, ldb
+    );
+}
+
+
+template <typename T>
+std::ostream &operator<<(std::ostream &os, std::vector<T> &v) {
+    size_t n = v.size();
+    os << "{";
+    if (n)
+    {
+        os << v[0];
+        for (size_t i = 1; i < n; ++i)
+            os << ", " << v[i];
+    }
+    os << "}";
+    return os;
+}
+
+
+} // end namespace RandBLAS_Testing::Util
 #endif
