+++ test/CMakeLists.txt
@@ -6,8 +6,14 @@ if (GTest_FOUND)
     set(tmp TRUE)
 
     add_executable(RandBLAS_tests
-        test_sparse.cc
-        test_dense.cc
+        # dense
+        test_dense/test_moments.cc
+        test_dense/test_sketch_gefl3.cc
+        test_dense/test_sketch_gefr3.cc
+        # sparse
+        test_sparse/test_construction.cc
+        test_sparse/test_sketch_gefls.cc
+        test_sparse/test_sketch_gefrs.cc
     )
     target_link_libraries(RandBLAS_tests
         RandBLAS
@@ -19,5 +25,5 @@ if (GTest_FOUND)
 endif()
 message(STATUS "Checking for regression tests ... ${tmp}")
 
-add_executable(test_rng_speed test_rng_speed.cpp)
+add_executable(test_rng_speed test_rng_speed.cc)
 target_link_libraries(test_rng_speed RandBLAS)
