+++ test/test_rng_speed.cc
